pyproject.toml
[tool.poetry]
name = "projeto-refatorado-etl-pandas"
version = "0.1.0"
description = ""
authors = ["Guilherme de Souza Melo"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pandas = "^2.2.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
